Consolidated function for term search
